Upgrader, tails-security-check: drop CA pinning
We don't manage to foresee CA changes and keep this up-to-date before it harms
our users. This provides little extra security, and ironically, when we fail to
update the CA, we expose our users to the very same practical effects as the
attack this mechanism is meant to protect against.

For details, see
https://gitlab.tails.boum.org/tails/tails/-/issues/18324#note_169820

Closes #18324