Test suite: drop Seahorse tests
These tests are the only part of our test suite that depends on v2
Onion Services.

Given we decided to stop investing much time into the "OpenPGP outside of
Thunderbird" area, in order to focus on the needs of our main personas, let's
remove these tests instead of porting them (which might be cheap, or not, but
even investigating that would be a poor use of our resources IMO).

refs #18256