Upgrader: ensure we fail closed on download and verification errors
I am quite uncomfortable with relying on the fact we will always do

   $success = $download_h->finish

… on any kind of download failure, knowing that a "download" failure can mean
"the checksum of the fully downloaded IUK is not the expected one" in this
context, so this could be a pretty severe security issue.

So instead, set $success to 0 by default, and set it to 1 only when we've
confirmed a successful (and verified) download.

refs #18162