Release process: ask clearer commitments from manual testers
I've seen cases when 1 of the 2 registered manual testers had not
actually blocked any time for the tests, and instead did 1 single
test while working on something non-Tails. That's totally OK,
but when that's what a manual tester means when they register,
as a RM I want to know in advance, so I can ask more people to register.