Test suite: remove GnuPG keyserver test
Most of our users will use other tools to fetch OpenPGP keys: Thunderbird,
Seahorse. These tools have their own implementation and don't use GnuPG /
dirmngr. On top of that, this test was not even using the keyserver configured
in Tails (because we can't use Onion services in our test suite), so what it was
exercising was not really close to a real Tails.

refs #18103