drop tca watchdog
it is not strictly needed. the rationale behind it was to be really sure
that we notice when tca exits, so that we can stop its dependencies.
But doing that doesn't require the watchdog to be present, closing the
process is enough for systemd to tear down dependencies. So let's remove
a moving part and make it simpler to understand.