Test suite: rework translate() to not always use the tails domain (fixes: #18447).
We currently have a localization test for Chinese that fails because
it wants to find "Thunderbird" using Dogtail, but in tails translation
domain for Chinese there is a translation "Thunderbird" →
"ThunderBird", which doesn't match.

So, now that our Dogtail integration always use translate() when
searching for nodes, let's make it opt-in, because it is only for TCA
that we currently need this, and it causes trouble elsewhere.