Doc: remove reference to .packages file
As pointed out on #18424, the contents of that file is not 100% correct.
I suppose most of our users would not be able to make much use of that
file, and those who do need something like this, such as the reporter
of #18424, will be better served by looking at the output of `dpkg -l`
inside Tails.