Test suite: work around the desktop icons being obscured by TCA.
If the network is plugged before this step we have a race where TCA
could start before we have found the desktop icon. With the workaround
from this commit the following scenario, that satisfies the above
condition, works fine:

  Scenario: test
    Given I have started Tails from DVD without network and logged in
    When the network is plugged
    And the Tor Connection Assistant autostarts
    Then the Tails desktop is ready

It could be that this workaround can be dropped if #18407 is fixed.