Fix feature branch commit encoded into the artifacts' names (fixes: #18468)
Since !381 (Merge base branch earlier) or some of its follow-ups we've
encoded the merge-commit as the feature branch's commit. This is a
problem since this commit often (and always on Jenkins) is throw-away
and thus not visible outside of the Git repo it was made in, making
this information useless. So let's encode the commit prior to the base
branch merge.